Refactor privacy group tests
