+++ acceptance-tests/src/test/java/tech/pegasys/pantheon/tests/acceptance/dsl/privacy/condition/PrivateContractCondition.java
@@ -10,11 +10,10 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.tests.acceptance.dsl.privacy.transaction;
+package tech.pegasys.pantheon.tests.acceptance.dsl.privacy.condition;
 
-import tech.pegasys.pantheon.tests.acceptance.dsl.privacy.PrivacyNode;
+import org.web3j.tx.Contract;
 
-public interface PrivateCondition {
-
-  void verify(PrivacyNode node);
+public interface PrivateContractCondition {
+  void verify(final Contract contract);
 }
